Add more constructors to DriveDistanceProfileCommand
These just give more options when starting a command, as we don't
want to always specify every value. Usually, we will just specify a distance
and the rest of the things will be the same, so there is a constructor for
that, along with a few others that may be useful.