2 ball Auton work
This contains a lot of work on the basic autons ranging to our attempt
of 2 ball auton.  It is a huge change due to conflicting issues in the
motion-profiling branch where it does not allow changes to be pulled.
The basic static autons were refined, and if 2 ball were to be attempted
,DriveInchesPIDController would have to be increased to full speed to
fit under the time slot. To properly continue good coding practices,
it is recommended that all these branches and conflicting commits all be
merged into master.